Release v0.13.0

+++ rpc/core/consensus.go
@@ -42,7 +42,7 @@ import (
 // 	"jsonrpc": "2.0"
 // }
 // ```
-func Validators(heightPtr *int) (*ctypes.ResultValidators, error) {
+func Validators(heightPtr *int64) (*ctypes.ResultValidators, error) {
 	if heightPtr == nil {
 		blockHeight, validators := consensusState.GetValidators()
 		return &ctypes.ResultValidators{blockHeight, validators}, nil
